move some tests to orders_test.go
